+++ butterfly-backup-web/bbweb/settings.py
@@ -11,9 +11,10 @@ https://docs.djangoproject.com/en/4.2/ref/settings/
 """
 
 from pathlib import Path
+from os import environ
 
 # Butterfly Backup catalog
-CATALOG_PATH = "/tmp/backup"
+CATALOG_PATH = environ.get("BB_CATALOG_PATH", "/tmp/backup")
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
@@ -127,3 +128,6 @@ STATIC_URL = "static/"
 # https://docs.djangoproject.com/en/4.2/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+SESSION_EXPIRE_AT_BROWSER_CLOSE = True
+SESSION_COOKIE_AGE = 180
+SESSION_SAVE_EVERY_REQUEST = True