+++ butterfly-backup-web/bbweb/urls.py
@@ -24,4 +24,6 @@ urlpatterns = [
     path("accounts/", include("django.contrib.auth.urls")),
     path("", views.home, name="home"),
     path("details/<str:section>", views.details, name="details"),
+    path("details/<str:section>/logs", views.logs, name="logs"),
+    path("backup/", views.backup, name="backup")
 ]