+++ CHANGES.md
@@ -9,13 +9,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
-## [0.0.1] - 1970-01-01
+## [0.0.1] - 2025-06-26
 
 ### Added
 - Start **butterfly-backup-web** project
+- Add homepage (catalog)
 
 ### Changed
 - None
+
+### Removed
+- None
+
+### Fixed
+- None
+
+## [0.0.2] - 2025-07-08
+
+### Added
+- Add details page
+- Add logs page
+- Add backup page
+- Add `BB_CATALOG_PATH` env variable
+
+### Changed
+- Add check into homepage if catalog exists
 
 ### Removed
 - None